+++ src/super/i-static-page/modules/ss-helpers/favicons.js
@@ -11,7 +11,7 @@ const
 
 const
 	glob = require('glob'),
-	fs = require('fs-extra-promise'),
+	fs = require('fs-extra'),
 	path = require('upath');
 
 const